Renamed package to customerlist
